drf/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.8
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-js-asset==2.2.0
django-jsoneditor==0.2.4
django-jsonform==2.22.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
executing==2.0.1
factory-boy==3.3.0
Faker==26.1.0
gunicorn==23.0.0
httpie==3.2.3
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kavenegar==1.1.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.0
rpds-py==0.19.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
